flex/api-flex-template/requirements.txt
apache-beam[gcp]==2.64.0
google-cloud-secret-manager==2.22.1
google-cloud-bigquery==3.30.0
jsonpath-python==1.0.6
requests==2.32.3
